Added new commands & Added CommandOutputBuilder class
- Added EntityInfoCommand
- Added ChunkInfoCommand
- Added WorldSizeCommand
- Added PerformanceSummaryCommand
- Created CommandOutputBuilder to make command output formatting more streamlined

+++ src/main/java/com/danieltebor/mc_server_analytics/util/MemInfo.java
@@ -29,7 +29,7 @@ import java.lang.management.MemoryUsage;
  * @author Daniel Tebor
  */
 public abstract class MemInfo {
-
+    
     public static long getCommittedHeapMemory() {
         return ManagementFactory.getMemoryMXBean().getHeapMemoryUsage().getCommitted();
     }
@@ -64,7 +64,7 @@ public abstract class MemInfo {
         return bytes / (1024 * 1024);
     }
 
-    public static long toGB(long bytes) {
+    public static float toGB(long bytes) {
         return bytes / (1024 * 1024 * 1024);
     }
 }
